Changes to check for constraint C1140 which prohibits deallocation of polymorphic entities in a DO CONCURRENT.  Deallocation can be caused by exiting from a block where the entity is declared, from an assignment, and from direct deallocation.
I added a function to tools.cc to check if a symbol is polymorphic.

I also added a function to check-do.cc to report errors along with a reference
to the declaration of the entity associated with error.  This function allows
for the possibility that the declaration of the entity is in a module.